add migration

+++ Basketball.League.Domain/Entities/City.cs
@@ -6,7 +6,7 @@ public class City
 {
     public int Id { get; set; }
     public string CityName { get; set; } = null!;
-    public BigInteger Population { get; set; }
+    public long Population { get; set; }
     public int StateId { get; set; }
 
     public State State { get; set; } = null!;
